refactor: extract handler logic to service files (handler-size-limit)
Decompose 10 oversized handlers into thin orchestrators by extracting
business logic to dedicated service files. All handlers now under 80
lines. Zero mantle check violations at all severities.

New service files: download/*, notification/*, apiGateway/*, auth/*,
device/*, cleanup/*, file/*, user/*

Confidence: high
Scope-risk: moderate
